+++ bpm-engine-impl/src/main/java/io/takari/bpm/commands/ClearCommandStackCommand.java
@@ -4,18 +4,8 @@ public class ClearCommandStackCommand implements Command {
 
     private static final long serialVersionUID = 1L;
 
-    private final String definitionId;
-
-    public ClearCommandStackCommand(String definitionId) {
-        this.definitionId = definitionId;
-    }
-
-    public String getDefinitionId() {
-        return definitionId;
-    }
-
     @Override
     public String toString() {
-        return "ClearCommandStackCommand [definitionId=" + definitionId + "]";
+        return "ClearCommandStackCommand []";
     }
 }